Creating new database
In this file, I have created a Database using MySQL, then added a table, learn how to use CHAR and VARCHAR to set the length of data, NULL or NOT NULL to set if the data is required or not, alongside the use of primary and foreign keys to create data and reuse date from one table to another. And finally the use of int (integer for numbers) and date for all kinds of dates.